#57 [feature] Main Activity data 받기, 파일 위치 refactor
