+++ app/src/main/java/com/pns/albang/view/MainActivity.kt
@@ -19,7 +19,6 @@ import com.naver.maps.map.overlay.Marker
 import com.naver.maps.map.overlay.OverlayImage
 import com.naver.maps.map.util.FusedLocationSource
 import com.pns.albang.R
-import com.pns.albang.ReviewActivity
 import com.pns.albang.TestActivity
 import com.pns.albang.databinding.ActivityMainBinding
 import com.pns.albang.viewmodel.MainViewModel
@@ -34,7 +33,7 @@ class MainActivity : AppCompatActivity() {
     private lateinit var circleOverlay: CircleOverlay
     private val landmarkInfoWindow = InfoWindow()
 
-    private var guestBookIntent = Intent(this, TestActivity::class.java)
+    private lateinit var guestBookIntent: Intent
 
     private val onMapReadyCallback = OnMapReadyCallback {
         this.naverMap = it.apply {
@@ -99,6 +98,8 @@ class MainActivity : AppCompatActivity() {
 
         fusedLocationSource = FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE)
 
+        guestBookIntent = Intent(this, TestActivity::class.java)
+
         val mapFragment = supportFragmentManager.findFragmentById(R.id.fragment_map) as MapFragment
         mapFragment.getMapAsync(onMapReadyCallback)
 
@@ -145,11 +146,7 @@ class MainActivity : AppCompatActivity() {
                     text = getString(R.string.btn_request)
                 }
             } else {
-                guestBookIntent = Intent(this@MainActivity, ReviewActivity::class.java).apply {
-                    putExtra("landmark", landmark.id)
-                    putExtra("landmarkImage", landmark.imageName)
-                    putExtra("landmarkName", landmark.name)
-                }
+                guestBookIntent = ReviewActivity.newIntent(this@MainActivity, landmark.id, landmark.imageName, landmark.name)
 
                 binding.btnBottom.apply {
                     icon = ContextCompat.getDrawable(this@MainActivity, R.drawable.ic_guestbook)
